Apply display precision of zero decimals
When CT's are installed there is more precision available, but that
has little added value. Also the other readings are integers, so it
looks nicer if the suggested display precision is 0.

Users can still manually override to show more precision as this is
just the default. The precision is still available if users want this.